requirements.txt
requests
bs4
langchain
langchain-huggingface
langchain-text-splitters
langchain-community
pinecone
python-dotenv
sentence-transformers
huggingface-hub
langchain-groq
gradio
fastapi
uvicorn
pydantic
ipykernel
langchain-experimental
emoji
matplotlib
pandas
numpy
langchain-pinecone
